refactor: split dashboard rails and settings logic
