+++ libs/workspace/dashboard/feature/src/index.ts
@@ -1,2 +1,6 @@
 export * from './lib/rails/workspace-dashboard-rails.component';
 export * from './lib/rails/dashboard-rail.component';
+export * from './lib/rails/dashboard-hero.utils';
+export * from './lib/rails/dashboard-live-epg.utils';
+export * from './lib/rails/dashboard-playback.utils';
+export * from './lib/rails/dashboard-rail.utils';
